+++ maintenance/controller/interfaces.go
@@ -2,21 +2,10 @@ package controller
 
 import (
 	"context"
-	libconfig "github.com/cloudogu/k8s-registry-lib/config"
+
 	"github.com/cloudogu/k8s-registry-lib/repository"
-	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
-type GlobalConfigRepository interface {
-	Get(context.Context) (libconfig.GlobalConfig, error)
-	Watch(context.Context, ...libconfig.WatchFilter) (<-chan repository.GlobalConfigWatchResult, error)
-	Update(ctx context.Context, globalConfig libconfig.GlobalConfig) (libconfig.GlobalConfig, error)
-}
-
-// used for mocks
-
-//nolint:unused
-//goland:noinspection GoUnusedType
-type k8sClient interface {
-	client.Client
+type MaintenanceAdapter interface {
+	GetStatus(ctx context.Context) (repository.MaintenanceModeDescription, bool, error)
 }