+++ maintenance/controller/maintenanceController.go
@@ -4,31 +4,21 @@ import (
 	"context"
 	"fmt"
 	"html/template"
-	"k8s.io/apimachinery/pkg/util/yaml"
-	"log"
 	"os"
 
+	"github.com/cloudogu/k8s-registry-lib/repository"
 	corev1 "k8s.io/api/core/v1"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/builder"
 	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/event"
 	"sigs.k8s.io/controller-runtime/pkg/predicate"
-
-	"encoding/json"
 )
 
 const (
-	maintenanceKey = "maintenance"
-	path503        = "/var/www/html/errors/503.html"
+	path503 = "/var/www/html/errors/503.html"
 )
 
 type TmplConfig map[string]string
-
-type MaintenanceMode struct {
-	Title string `json:"title,omitempty"`
-	Text  string `json:"text,omitempty"`
-}
 
 func (c TmplConfig) GetOrDefault(key, def string) string {
 	if v, ok := c[key]; ok && v != "" {
@@ -41,52 +31,40 @@ type PageData struct {
 	Config TmplConfig
 }
 
-// MaintenanceReconciler is responsible for reconciling the global configmap and to create a corresponding error page
+// MaintenanceReconciler is responsible for reconciling the maintenance configmap and to create a corresponding error page
 // for the maintenance mode
 type MaintenanceReconciler struct {
-	Client             client.Client
-	GlobalConfigGetter GlobalConfigRepository
+	Adapter MaintenanceAdapter
 }
 
-// Reconcile reconciles the global configmap and triggers the error page generation
-func (r *MaintenanceReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Result, error) {
+// Reconcile reconciles the maintenance configmap and triggers the error page generation
+func (r *MaintenanceReconciler) Reconcile(ctx context.Context, _ ctrl.Request) (ctrl.Result, error) {
 	logger := ctrl.LoggerFrom(ctx)
-	logger.Info("Reconciling global config for redirect")
+	logger.Info("Reconciling maintenance config for redirect")
 
-	cm := &corev1.ConfigMap{}
-	err := r.Client.Get(ctx, req.NamespacedName, cm)
+	description, active, err := r.Adapter.GetStatus(ctx)
 	if err != nil {
-		return ctrl.Result{}, fmt.Errorf("failed to get global config map: %w", err)
+		return ctrl.Result{}, fmt.Errorf("failed to get maintenance status: %w", err)
 	}
 
-	globalCfg, err := r.GlobalConfigGetter.Get(ctx)
-	if err != nil {
-		return ctrl.Result{}, fmt.Errorf("failed to get global config: %w", err)
-	}
-
-	maintenanceModeJson, ok := globalCfg.Get(maintenanceKey)
-	if !ok {
-		return ctrl.Result{}, fmt.Errorf("maintenance not found in global config")
-	}
-	logger.Info(fmt.Sprintf("Mode: %s", maintenanceModeJson))
-
-	maintenanceMode := &MaintenanceMode{}
-	err = json.Unmarshal([]byte(maintenanceModeJson), maintenanceMode)
-	if err != nil {
-		return ctrl.Result{}, fmt.Errorf("failed to parse maintenancemode: %w", err)
+	if !active {
+		logger.Info("Maintenance mode is inactive. Skipping maintenance page generation.")
+		return ctrl.Result{}, nil
 	}
 
 	tmpl := template.Must(template.ParseFiles(fmt.Sprintf("%s.tpl", path503)))
 
 	data := PageData{
 		Config: TmplConfig{
-			"maintenance/title": maintenanceMode.Title,
-			"maintenance/text":  maintenanceMode.Text,
+			"maintenance/title": description.Title,
+			"maintenance/text":  description.Text,
 		},
 	}
 	if err = renderToFile(tmpl, path503, data); err != nil {
-		log.Fatal(err)
+		logger.Error(err, "Failed to render maintenance page")
+		return ctrl.Result{}, fmt.Errorf("failed to render maintenance page: %w", err)
 	}
+
 	return ctrl.Result{}, nil
 }
 
@@ -101,54 +79,16 @@ func renderToFile(t *template.Template, outPath string, data any) error {
 	return t.Execute(f, data)
 }
 
-// SetupWithManager sets up the global configmap controller with the Manager.
-// The controller watches for changes to the global configmap.
+// SetupWithManager sets up the maintenance configmap controller with the Manager.
+// The controller watches for changes to the maintenance configmap.
 func (r *MaintenanceReconciler) SetupWithManager(mgr ctrl.Manager) error {
 	return ctrl.NewControllerManagedBy(mgr).
-		For(&corev1.ConfigMap{}, builder.WithPredicates(globalConfigPredicate())).
-		WithEventFilter(MaintenanceChangedPredicate()).
+		For(&corev1.ConfigMap{}, builder.WithPredicates(maintenancePredicate())).
 		Complete(r)
 }
 
-func globalConfigPredicate() predicate.Funcs {
+func maintenancePredicate() predicate.Funcs {
 	return predicate.NewPredicateFuncs(func(object client.Object) bool {
-		return object.GetName() == "global-config"
+		return object.GetName() == repository.MaintenanceConfigMapName
 	})
-}
-
-type Config struct {
-	Maintenance string `yaml:"maintenance"`
-}
-
-func MaintenanceChangedPredicate() predicate.Predicate {
-	return predicate.Funcs{
-		UpdateFunc: func(e event.UpdateEvent) bool {
-			oldCM, ok1 := e.ObjectOld.(*corev1.ConfigMap)
-			newCM, ok2 := e.ObjectNew.(*corev1.ConfigMap)
-			if !ok1 || !ok2 {
-				return false
-			}
-
-			oldCfg := &Config{}
-			newCfg := &Config{}
-
-			if dataOld, ok := oldCM.Data["config.yaml"]; ok {
-				_ = yaml.Unmarshal([]byte(dataOld), oldCfg)
-			}
-			if dataNew, ok := newCM.Data["config.yaml"]; ok {
-				_ = yaml.Unmarshal([]byte(dataNew), newCfg)
-			}
-
-			return oldCfg.Maintenance != newCfg.Maintenance
-		},
-		CreateFunc: func(e event.CreateEvent) bool {
-			return true
-		},
-		DeleteFunc: func(e event.DeleteEvent) bool {
-			return true
-		},
-		GenericFunc: func(e event.GenericEvent) bool {
-			return false
-		},
-	}
 }