+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- [#15] Use the new maintenance ConfigMap
+  - Previously, the maintenance mode was read from the global config
 
 ## [v1.0.4] - 2025-11-27
 ### Changed